+++ examples/hello-world/app/app.js
@@ -1,5 +1,11 @@
 // Copyright Jon Williams 2017. See LICENSE file.
 import React from "react";
+import Title from "app/title";
 
 /* A simple (but friendly) React Component */
-export default () => <h1>Hello World!</h1>;
+export default () => (
+  <div>
+    <Title>Hello Down There!</Title>
+    <h1>Hello World!</h1>
+  </div>
+);