+++ examples/hello-world/index.test.js
@@ -8,6 +8,11 @@ describe("Hello-World Integration Test", () => {
   test("should be polite and say Hello World", async () => {
     const res = await request(handler).get("/");
     expect(res.statusCode).toBe(200);
-    expect(res.text).toEqual(expect.stringContaining("Hello World"));
+
+    // The main text
+    expect(res.text).toEqual(expect.stringContaining("Hello World!"));
+    // Text in the title
+    expect(res.text).toEqual(expect.stringContaining("<title>Hello Down There!</title>"));
   });
+
 });