+++ packages/cater-build/server/cater/skipserversiderender.js
@@ -3,6 +3,4 @@ import React from "react";
 
 // This is the server-side version of this skip component - so return no
 // result.
-export default () => {
-  return false;
-};
+export default () => false;