+++ packages/cater-runtime/src/cater-context.js
@@ -9,16 +9,23 @@ class CaterContext {
   constructor(bundlePath) {
     this.bundlePath = bundlePath;
     this.globalStyles = [];
+    this.title = "";
   }
 
   /**
    * Adds a stylesheet that is intended to be rendered in the head of the
    * component. See the GlobalStyles component, which uses this function.
-   *
-   * @param {*} href
    */
   addGlobalStyle(href) {
     this.globalStyles.push(href);
+  }
+
+  /**
+   * Sets the title for this page. Only the last title will apply if this
+   * is called multiple times in a single render.
+   */
+  setTitle(title) {
+    this.title = title;
   }
 
 }