+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- app/title Title component will set the server and client side title tag in the head of the document.
+
 ## [0.3.6] - 2017-12-29
 ### Added
 - A config.cater.js file in the application root directory now is used as options. This can include an "env" section that has options specific for a NODE_ENV.